+++ packages/@react-elf/ui/src/index.ts
@@ -7,3 +7,4 @@ export * from "@react-elf/menu";
 export * from "@react-elf/divider";
 export * from "@react-elf/popover";
 export * from "@react-elf/notification";
+export * from "@react-elf/dialog";